requirements.txt
mcp>=1.0.0
twilio>=9.0.0
aiohttp>=3.9.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
tzdata>=2024.1
